pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-bootstrap5>=25.1",
    "flake8>=7.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
]

[build-system]
requires = [
    "setuptools>=80.7.1",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
